fix: don't use the deno cache for data urls
